Refactor - Added custom hook to get window dimensions and code cleanup
