+++ src/data/imageIds.js
@@ -1,0 +1,9 @@
+const imageIds = [
+  11, 16, 30, 52, 53, 110, 114, 116, 117, 120, 124, 128, 131, 135, 149, 162,
+  172, 173, 175, 184, 185, 194, 249, 271, 273, 279, 307, 327, 329, 377, 425,
+  445, 446, 450, 499, 500, 505, 516, 525, 536, 537, 541, 544, 545, 560, 563,
+  564, 570, 594, 654, 681, 700, 719, 721, 723, 765, 791, 1032, 1033, 1038, 1049,
+  1055, 1068, 1083, 1500,
+];
+
+export default imageIds;
